+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix sku autocomplete section selected item
+- Updated the create wishlist modal
+- Fix adding multiple sku issue.
 
 ## [1.0.12] - 2024-05-09